refactor(ci): extract SARIF upload slice into phi_scan/ci/sarif.py
Moves upload_sarif_to_github and its 3 helpers from ci_integration.py into a dedicated submodule. All symbols remain importable from phi_scan.ci_integration via re-export. Locks surface contract via new parity test.
